Struct imap::client::IMAPStream

pub struct IMAPStream {
    // some fields omitted
}

Stream to interface with the IMAP server. This interface is only for the command stream.

Methods

impl IMAPStream

fn connect<A: ToSocketAddrs>(addr: A, ssl_context: Option<SslContext>) -> Result<IMAPStream>

Creates an IMAP Stream.

fn login(&mut self, username: &str, password: &str) -> Result<()>

Log in to the IMAP server.

fn select(&mut self, mailbox_name: &str) -> Result<IMAPMailbox>

Selects a mailbox

fn examine(&mut self, mailbox_name: &str) -> Result<IMAPMailbox>

Examine is identical to Select, but the selected mailbox is identified as read-only

fn fetch(&mut self, sequence_set: &str, query: &str) -> Result<Vec<String>>

Fetch retreives data associated with a message in the mailbox.

fn noop(&mut self) -> Result<()>

Noop always succeeds, and it does nothing.

fn logout(&mut self) -> Result<()>

Logout informs the server that the client is done with the connection.

fn create(&mut self, mailbox_name: &str) -> Result<()>

Create creates a mailbox with the given name.

fn delete(&mut self, mailbox_name: &str) -> Result<()>

Delete permanently removes the mailbox with the given name.

fn rename(&mut self, current_mailbox_name: &str, new_mailbox_name: &str) -> Result<()>

Rename changes the name of a mailbox.

fn subscribe(&mut self, mailbox: &str) -> Result<()>

Subscribe adds the specified mailbox name to the server's set of "active" or "subscribed" mailboxes as returned by the LSUB command.

fn unsubscribe(&mut self, mailbox: &str) -> Result<()>

Unsubscribe removes the specified mailbox name from the server's set of "active" or "subscribed" mailboxes as returned by the LSUB command.

fn capability(&mut self) -> Result<Vec<String>>

Capability requests a listing of capabilities that the server supports.

fn expunge(&mut self) -> Result<()>

Expunge permanently removes all messages that have the \Deleted flag set from the currently selected mailbox.

fn check(&mut self) -> Result<()>

Check requests a checkpoint of the currently selected mailbox.

fn close(&mut self) -> Result<()>

Close permanently removes all messages that have the \Deleted flag set from the currently selected mailbox, and returns to the authenticated state from the selected state.

fn copy(&mut self, sequence_set: &str, mailbox_name: &str) -> Result<()>

Copy copies the specified message to the end of the specified destination mailbox.

fn run_command_and_check_ok(&mut self, command: &str) -> Result<()>

fn run_command(&mut self, untagged_command: &str) -> Result<Vec<String>>