Struct imap::client::IdleHandle

pub struct IdleHandle<'a, T: Read + Write + 'a> { /* fields omitted */ }

IdleHandle allows a client to block waiting for changes to the remote mailbox.

The handle blocks using the IMAP IDLE command specificed in RFC 2177.

As long a the handle is active, the mailbox cannot be otherwise accessed.

Methods

impl<'a, T: Read + Write + 'a> IdleHandle<'a, T>

fn wait(&mut self) -> Result<()>

Block until the selected mailbox changes.

fn cancel(self)

Cancel the IDLE handle prematurely.

impl<'a, T: SetReadTimeout + Read + Write + 'a> IdleHandle<'a, T>

fn set_keepalive(&mut self, interval: Duration)

Set the keep-alive interval to use when wait_keepalive is called.

The interval defaults to 29 minutes as dictated by RFC 2177.

fn wait_keepalive(&mut self) -> Result<()>

Block until the selected mailbox changes.

This method differs from IdleHandle::wait in that it will periodically refresh the IDLE connection, to prevent the server from timing out our connection. The keepalive interval is set to 29 minutes by default, as dictated by RFC 2177, but can be changed using set_keepalive.

This is the recommended method to use for waiting.

fn wait_timeout(&mut self, timeout: Duration) -> Result<()>

Block until the selected mailbox changes, or until the given amount of time has expired.