imap::types

Enum UnsolicitedResponse

pub enum UnsolicitedResponse {
    Status {
        mailbox: String,
        attributes: Vec<StatusAttribute>,
    },
    Recent(u32),
    Exists(u32),
    Expunge(Seq),
}

Responses that the server sends that are not related to the current command. RFC 3501 states that clients need to be able to accept any response at any time. These are the ones we’ve encountered in the wild.

Note that Recent, Exists and Expunge responses refer to the currently SELECTed folder, so the user must take care when interpreting these.

Variants

Status

An unsolicited STATUS response.

Fields

mailbox: String

The mailbox that this status response is for.

attributes: Vec<StatusAttribute>

The attributes of this mailbox.

Recent(u32)

An unsolicited RECENT response indicating the number of messages with the \Recent flag set. This response occurs if the size of the mailbox changes (e.g., new messages arrive).

Note: It is not guaranteed that the message sequence numbers of recent messages will be a contiguous range of the highest n messages in the mailbox (where n is the value reported by the RECENT response). Examples of situations in which this is not the case are: multiple clients having the same mailbox open (the first session to be notified will see it as recent, others will probably see it as non-recent), and when the mailbox is re-ordered by a non-IMAP agent.

The only reliable way to identify recent messages is to look at message flags to see which have the \Recent flag set, or to do a SEARCH RECENT.

Exists(u32)

An unsolicited EXISTS response that reports the number of messages in the mailbox. This response occurs if the size of the mailbox changes (e.g., new messages arrive).

Expunge(Seq)

An unsolicited EXPUNGE response that reports that the specified message sequence number has been permanently removed from the mailbox. The message sequence number for each successive message in the mailbox is immediately decremented by 1, and this decrement is reflected in message sequence numbers in subsequent responses (including other untagged EXPUNGE responses).

The EXPUNGE response also decrements the number of messages in the mailbox; it is not necessary to send an EXISTS response with the new value.

As a result of the immediate decrement rule, message sequence numbers that appear in a set of successive EXPUNGE responses depend upon whether the messages are removed starting from lower numbers to higher numbers, or from higher numbers to lower numbers. For example, if the last 5 messages in a 9-message mailbox are expunged, a “lower to higher” server will send five untagged EXPUNGE responses for message sequence number 5, whereas a “higher to lower server” will send successive untagged EXPUNGE responses for message sequence numbers 9, 8, 7, 6, and 5.