imap::types

Struct Capabilities

pub struct Capabilities(/* private fields */);

From section 7.2.1 of RFC 3501.

A list of capabilities that the server supports. The capability list will include the atom “IMAP4rev1”.

In addition, all servers implement the STARTTLS, LOGINDISABLED, and AUTH=PLAIN (described in IMAP-TLS) capabilities. See the Security Considerations section of the RFC for important information.

A capability name which begins with AUTH= indicates that the server supports that particular authentication mechanism.

The LOGINDISABLED capability indicates that the LOGIN command is disabled, and that the server will respond with a super::Error::No response to any attempt to use the LOGIN command even if the user name and password are valid. An IMAP client MUST NOT issue the LOGIN command if the server advertises the LOGINDISABLED capability.

Other capability names indicate that the server supports an extension, revision, or amendment to the IMAP4rev1 protocol. Capability names either begin with X or they are standard or standards-track RFC 3501 extensions, revisions, or amendments registered with IANA.

Client implementations SHOULD NOT require any capability name other than IMAP4rev1, and MUST ignore any unknown capability names.

Implementations

impl Capabilities

pub fn has<'a>(&self, cap: &Capability<'a>) -> bool

Check if the server has the given capability.

pub fn has_str<S: AsRef<str>>(&self, cap: S) -> bool

Check if the server has the given capability via str.

pub fn iter(&self) -> Iter<'_, Capability<'_>>

Iterate over all the server’s capabilities

pub fn len(&self) -> usize

Returns how many capabilities the server has.

pub fn is_empty(&self) -> bool

Returns true if the server purports to have no capabilities.